Resolve dependency issues
